move reference implementation and rename folder
