--- mlbench/master/main/templates/main/test.html
@@ -1,46 +1,0 @@
-{% extends 'main/base.html' %}
-
-{% block content %}
-<div class="row" id="results">
-
-</div>
-<div class="row">
-    <button type="button" class="btn btn-primary" id="start-test-button">Start Test</button>
-</div>
-{% endblock %}
-{% block scripts %}
-<script id="result-template" type="text/x-jsrender">
-    <ul class="list-group">
-        <div class="d-flex w-100 justify-content-between">
-            <span class="mb-1">Nodes: </span>
-            <span><%>nodes%></span>
-        </div>
-        <div class="d-flex w-100 justify-content-between">
-            <span class="mb-1">Command: </span>
-            <span><%>command%></span>
-        </div>
-        <div class="d-flex w-100 justify-content-start">
-            <span class="mb-1">Master: </span>
-            <span><%>master_name%></span>
-        </div>
-        <div class="d-flex w-100 justify-content-start">
-            <span class="mb-1">Response: </span>
-            <span style="white-space:pre-wrap;"><pre><%>response%></pre></span>
-        </div>
-    </ul>
-</script>
-<script type="text/javascript">
-    $(document).ready(function () {
-        $.views.settings.delimiters("<%", "%>");
-
-        $("#start-test-button").click(function(){
-            $("#results").empty();
-
-            $.post( "api/mpi_jobs/", function( data ) {
-                var res = $("#result-template").render(data);
-                $("#results").append(res)
-            });
-        })
-    })
-</script>
-{% endblock %}
