+++ mlbench/master/main/views.py
@@ -25,7 +25,7 @@ def run(request, run_id):
 
     run.job_metadata = job.meta
 
-    print(run.job_metadata)
+    run.job_metadata['stdout'] = "\n".join(run.job_metadata['stdout'])
     run.job_metadata['stderr'] = "\n".join(run.job_metadata['stderr'])
 
     return render(request,
