--- refimpls/Makefile
@@ -1,2 +1,0 @@
-doc:
-	sphinx-apidoc -o docs/ codes/
